Changes in project structure. New folder for Test-related files.
