+++ UnitTests/TestingCellClass.cs
@@ -2,10 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 using NUnit.Framework;
-using System.Threading;
 using System.IO;
 
-namespace GameLifeCSharpConsole
+namespace GameLifeCSharpConsole.UnitTests
 {
     [TestFixture]
     class TestingCellClass
@@ -24,7 +23,7 @@ namespace GameLifeCSharpConsole
             return testCellDead;
         }
 
-        [Test]        
+        [Test]
         public void ChangeGenerationTest()
         {
             Cell testCellActive = new Cell(true);
@@ -66,69 +65,4 @@ namespace GameLifeCSharpConsole
             Assert.IsFalse(GetInactiveCellSample().IsActive && (activeCells < 2 || activeCells > 3));
         }
     }
-
-    [TestFixture]
-    class TestingFieldClass
-    {
-        //[Test]
-        
-    }
-
-    [TestFixture]
-    class TestingExceptions
-    {
-        [Test]             
-        public void WrongInputFieldHeight()
-        {
-            Field field = new Field(20, 40, false);
-            string path = "inputWrongHeight.txt";
-            Assert.Throws<InvalidFieldHeightException>(() => field.Init(path));
-        }
-
-        [Test]
-        public void WrongInputFieldWidth()
-        {
-            Field field = new Field(20, 40, false);
-            string path = "inputWrongWidth.txt";
-            Assert.Throws<InvalidFieldWidthException>(() => field.Init(path));
-        }
-
-        [Test]
-        public void WrongInputFieldSymbols()
-        {
-            Field field = new Field(20, 40, false);
-            string path = "inputWrongSymbols.txt";
-            Assert.Throws<InvalidInputSymbolsException>(() => field.Init(path));           
-        }
-
-        [Test]
-        public void WrongInputFieldSAllZeroes()
-        {
-            Exception e = new Exception();
-            Field field = new Field(20, 40, false);
-            string path = "inputAllZeroes.txt";
-            field.Init(path);
-
-            bool hasAnyActiveCells = false;
-            for (int y = 0; y < field.Height; y++)
-            {
-                for (int x = 0; x < field.Width; x++)
-                {
-                    if (field.GetCell(y, x).IsActive || field.GetCell(y, x).WillBeActive)
-                    {
-                        hasAnyActiveCells = true;
-                    }
-                }
-            }
-
-            if (!hasAnyActiveCells)
-            {
-                e = new NoActiveCellsException();
-            }
-
-            Assert.AreEqual(e.Message, new NoActiveCellsException().Message);
-        }
-
-    }
-
 }
