+++ UnitTests/inputAllZeroes.txt
